Check and update CMakeLists in every dir
Before, when creating unt tests using scargo gen command, it was only modifying CMakeLists file in the parent directory of unit tests file (not touching folders above in path).

Right now, it is checking and updating CMakeLists file on every level of path between tests directory and unit tests file directory.

I am not familiar with embeded software and structure of C projects, please verify additionaly if it is not breaking anything.

https://github.com/Spyro-Soft/scargo/issues/134